requirements.txt
django-tinymce
django-admin-interface
django